lists schedule ok

+++ src/main/java/be/voedsaam/vzw/business/repository/impl/ScheduleServiceRepoImpl.java
@@ -1,19 +1,30 @@
 package be.voedsaam.vzw.business.repository.impl;
 
 import be.voedsaam.vzw.business.Schedule;
+import be.voedsaam.vzw.business.User;
 import be.voedsaam.vzw.business.repository.ScheduleRepository;
+import be.voedsaam.vzw.business.repository.UserRepository;
 import be.voedsaam.vzw.service.ScheduleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+
 @Service
 @Profile("springdatajpa")
 public class ScheduleServiceRepoImpl implements ScheduleService {
 
     private ScheduleRepository scheduleRepository;
+    private UserRepository userRepository;
+
+    @Autowired
+    public void setUserRepository(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
     @Autowired
     public void setScheduleRepository(ScheduleRepository scheduleRepository) {
         this.scheduleRepository = scheduleRepository;
@@ -37,7 +48,12 @@ public class ScheduleServiceRepoImpl implements ScheduleService {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws UnsupportedOperationException{
+        Schedule schedule = getById(id);
+        if (schedule.getUsers().size()!=0)
+            throw new UnsupportedOperationException("Schedule still has users please delete users from Schedule");
+        if (schedule.getDrives().size()!=0)
+            throw new UnsupportedOperationException("Schedule still has drives please delete users from Schedule");
         scheduleRepository.deleteById(id);
     }
 }
